getting all the workouts from the database to the tier 2

+++ ReSEP3-T3/src/main/java/resep3/javat3/resep3/javat3/protobuf/WorkoutPlanResponseOrBuilder.java
@@ -8,8 +8,8 @@ public interface WorkoutPlanResponseOrBuilder extends
     com.google.protobuf.MessageOrBuilder {
 
   /**
-   * <code>int64 wpid = 1;</code>
+   * <code>int32 wpid = 1;</code>
    * @return The wpid.
    */
-  long getWpid();
+  int getWpid();
 }
